Move render loop pausing to a separate class.
