+++ lib/ViewModels/TerriaViewer.js
@@ -166,8 +166,8 @@ function changeSelectBox(viewer) {
     } else {
         if (selectBox) {
             // Add and start a DrawExtentHelper - used by mapInteractionMode with drawRectangle
-            viewer._enableSelectExtent(terria.cesium.viewer.scene, false);
-            viewer.dragBox = new DrawExtentHelper(terria, terria.cesium.viewer.scene,
+            viewer._enableSelectExtent(terria.cesium.cesiumWidget.scene, false);
+            viewer.dragBox = new DrawExtentHelper(terria, terria.cesium.cesiumWidget.scene,
                 function(ext) {
                     var mapInteractionModeStack = this.terria.mapInteractionModeStack;
                     if (defined(mapInteractionModeStack) && mapInteractionModeStack.length > 0) {
@@ -179,7 +179,7 @@ function changeSelectBox(viewer) {
             viewer.dragBox.start();
         } else {
             viewer.dragBox.destroy();
-            viewer._enableSelectExtent(terria.cesium.viewer.scene, true);
+            viewer._enableSelectExtent(terria.cesium.cesiumWidget.scene, true);
         }
     }
 }
@@ -372,7 +372,7 @@ TerriaViewer.prototype.destroy = function () {
 };
 
 TerriaViewer.prototype.destroyCesium = function () {
-    const viewer = this.terria.cesium.viewer;
+    const cesiumWidget = this.terria.cesium.cesiumWidget;
 
     this.terria.cesium.destroy();
 
@@ -383,9 +383,9 @@ TerriaViewer.prototype.destroyCesium = function () {
 
     this.dataSourceDisplay.destroy();//
 
-    this._enableSelectExtent(viewer.scene, false);
+    this._enableSelectExtent(cesiumWidget.scene, false);
 
-    var inputHandler = viewer.screenSpaceEventHandler;
+    var inputHandler = cesiumWidget.screenSpaceEventHandler;
     inputHandler.removeInputAction(ScreenSpaceEventType.MOUSE_MOVE);
     inputHandler.removeInputAction(ScreenSpaceEventType.LEFT_DOUBLE_CLICK);
     inputHandler.removeInputAction(ScreenSpaceEventType.LEFT_DOUBLE_CLICK, KeyboardEventModifier.SHIFT);
@@ -394,7 +394,7 @@ TerriaViewer.prototype.destroyCesium = function () {
         this.monitor.destroy();
         this.monitor = undefined;
     }
-    viewer.destroy();
+    cesiumWidget.destroy();
     this.terria.cesium = undefined;
 };
 
@@ -582,7 +582,7 @@ that option from the Maps button at the top of the screen.'
 
 
         eventHelper = new EventHelper();
-        eventHelper.add(that.terria.cesium.viewer.clock.onTick, function(clock) {
+        eventHelper.add(that.terria.cesium.cesiumWidget.clock.onTick, function(clock) {
             that.dataSourceDisplay.update(clock.currentTime);
         });
         that.cesiumEventHelper = eventHelper;
